refactored project folders, made a dedicated folder for classes with derivations
