+++ .gitignore
@@ -30,6 +30,7 @@
 *.exe
 *.out
 *.app
+build.sh
 
 # dirs
 bin
