+++ include/board/attackboard.h
@@ -1,0 +1,8 @@
+#ifndef ATTACKBOARD_H
+#define ATTACKBOARD_H
+
+#include "board.h"
+
+class Attackboard : public Board {};
+
+#endif
