+++ include/board/defenseboard.h
@@ -1,0 +1,8 @@
+#ifndef DEFENSEBOARD_H
+#define DEFENSEBOARD_H
+
+#include "board.h"
+
+class Defenseboard : public Board {};
+
+#endif
