--- src/battleship.cpp
@@ -1,4 +1,0 @@
-//
-// Created by filip on 25/12/2022.
-//
-#include "battleship.h"
