+++ src/ship/battleship.cpp
@@ -1,0 +1,4 @@
+//
+// Created by filip on 25/12/2022.
+//
+#include "./../../include/ship/battleship.h"
